requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
decorator==5.1.1
flower==2.0.1
greenlet==3.0.3
humanize==4.9.0
idna==3.6
kafka-python==2.0.2
kombu==5.3.5
mysql-replication==1.0.6
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pika==1.3.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
py==1.11.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
retry==0.9.2
six==1.16.0
SQLAlchemy==2.0.27
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
